move base button to main components folder, rename to ArrowCircularButton
