+++ src/components/ArrowCircularButton/index.module.scss
@@ -1,4 +1,4 @@
-@use '../../../../../../styles/utilities/variables' as *;
+@use '../../styles//utilities/variables' as *;
 
 .root-detached {
   height: 100%;
